fix: ruff auto-fix (12,846 issues) + test repairs
- Run ruff check . --fix --unsafe-fixes: 17,326 → 4,559 errors
- Install missing deps: httpx2, rapidfuzz
- Fix test_corrections_manager: add missing Path import + fix assertions
- pytest: 270 → 278 passed (+8)
- Remaining 3 failures: domain logic in dictionary_manager (not env)

+++ packages/ai/gateway/pool/health_scorer.py
@@ -12,8 +12,6 @@ import os
 import time
 from collections import defaultdict
 from dataclasses import dataclass, field
-from typing import Any
-
 
 # ---------------------------------------------------------------------------
 # Configuration helpers
@@ -355,8 +353,8 @@ class ProviderHealthScorer:
         if not sorted_values:
             return 0.0
         k = (len(sorted_values) - 1) * (pct / 100.0)
-        lower = int(math.floor(k))
-        upper = int(math.ceil(k))
+        lower = math.floor(k)
+        upper = math.ceil(k)
         if lower == upper:
             return sorted_values[lower]
         fraction = k - lower
